+++ windowfunction.h
@@ -6,9 +6,9 @@
 
 class Windowfunction{
 public:
-    static void funcHammingTukey(std::vector<double> *winFunction);
-    static void funcHamming(std::vector<double> *winFunction, int size);
-    static void funcTukey(double *array, int size);
+    static void funcHammingTukey(std::vector<float> *winFunction);
+    static void funcHamming(std::vector<float> *winFunction, int size);
+    static void funcTukey(float *array, int size);
 };
 
 #endif // WINDOWFUNCTION_H
